requirements.txt
click>=8.1.8

# CrewAI
crewai==0.130.0
crewai-tools==0.47.1

# Frameworks
fastapi==0.110.3
uvicorn==0.29.0
starlette==0.37.2
anyio==4.3.0

# Google & Vertex AI
google-ai-generativelanguage==0.6.4
google-api-core==2.17.1
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.53.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0

# Jinja
Jinja2==3.1.4

# Validation & Utilities
pydantic==2.10.4
jsonschema==4.22.0

# Data & AI
numpy==1.26.4
pandas==2.2.2
protobuf==4.25.3

# LLM / Agents
openai==1.70.0
langchain-core==0.1.52
langsmith==0.1.67

# Networking stability fixes
httpx==0.27.0
httpcore==1.0.2

# Observability
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0

# Misc
oauthlib==3.2.2
pillow==10.3.0
PyPDF2==3.0.0
python-dotenv==1.0.0
